Implement collapsible categories in AppStore
Taking advantage of react-virtualized, we're hiding categories by
skipping the render method unless it's in `visibleChannels`.

`visibleChannels` is calculated by factoring in the categories that have
been hidden and removing the channels 'under it'. The category itself is
always shown.

This commit also includes functionality for saving/storing to
localStorage so that collapsed categories persist across browser
refreshes.

Resolves: #200