Extract emoji into a component + add click trap
The previous commit was more of a proof of concept. This commit builds
upon that and cleans up the code. I'm thinking it may be worth forking
emoji-picker-react as it seems 'stable'.

The reason to do so is to add guilds on the left side of the emoji
picker, a GIF selection menu, and stickers in the future.

This commit also adds a click trap so any clicks outside the emoji
picker dismiss the window.

Note: There's 2 potential ways we can hide the emoji picker:

1. Hide using `display: none`
2. Conditionally rendering the component (`conditional ? <Component /> :
null`)

I chose option 2 because the emoji picker is rarely used, no need to use
the memory.